Stuff
Moved ability-related things in the Kit class to the Ability class
